Few bug fixes
Same problem with loading of non existing libraries.
Changed the way the preprocessor macros are checked.
Now uses #if to check the value instead of the #if defined which checks if the actual value is defined

The PushLayer(Layer()) variant was not actually calling the onAttach function of the derived class. This now also uses a template to make sure the right derived class is returned.